Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I use three keyboards in rotation, a MBP keyboard, a 122-key Model M, and a standard 105 key, all ISO layout; therefore I'll just leave my two cents here:
> 
> - Having two command and two option keys on each side of the space key on the MBP keyboard is a bliss, it saves so many finger jumps just to type a shortcut or a special character.
> - Haiku at the moment is very confused with the MBP keyboard, this should be fixed.

I can’t tell what language these keyboards are in / you are using but if you are talking about French or some language that features AltGr on their PC keyboards, shouldn’t the (Mac) keymap already work this way in your language with the keys mirrored on the left and right on your mbp?

> - Having Alt Gr do something different than Alt will confuse people in "Haiku context". They should both function as Command (see first point about the niceties of two command keys)

In the French (Mac) keymap no, in the PC keymap yes, at least that’s what I’ve been lead to believe.

> - In addition to the above point, Windows and Linux people are already confused with the Haiku naming of Command/Option. Let's not confuse them more, and have an Alts and Windows keys groups doing only one thing.

We at least use Alt in the menus to not confuse PC users, and then in keymap we use the control, command, and option roles, and hope that people understand how roles translate to keys based on the visual representation.

> - BeOS clearly followed Mac OS convention in key naming, I don't think we should try to appease PC people. Instead they should get used to the Haiku way.

Yes this is because when BeOS ran on PPC Macs way back in the day using a clover symbol and OPT in menus, and then the menus keys were half updated after switching to PCs to say ALT or OPT... but this doesn’t really make much sense. It should either be ALT and WIN or CMD/clover and OPT but never ALT and OPT. Personally I think that putting CMD and OPT in the menus is a bad idea unless you are on an Apple keyboard (and we ascertain that info somehow) because “monkey see monkey do” is a powerful tool of human understanding. Ironically monkeys do not have the ability to see something and recognize it in other contexts, only human beings are able to do this, and only after they are ~3 years old.

> - Option (Windows/Super) key should continue doing what it's doing, either Stack, Tile or entering special characters.

Ug I agree with you, but apparently that’s not what users expect on keyboards with an AltGr key...

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>