Update patch set 1
Patch Set 1:

> No users won’t get used to, they won’t look it up in the user guide, they won’t look at Keymap preferences, they will simply be forever confused. We cannot ignore human nature when we design our HCI, that is folly.

If we solve this problem with AltGr I think we can manage to show the actual key names to users (control / alt / windows / altgr on pc; control/command/option on mac keyboards), actually matching their keyboards, in menus. So only the keymap preferences will remain a bit confusing by talking about command and option in all cases. But we can revisit (again) how we handle that.

There is the additional problem of having to handle cases where the PC and MAc keyboards are different (as is the case for French), and the cases where they are identical except for the swapping of command and option (as is the case for US).

But yes, Apple did mess things up when they designed their USB keyboards, I guess? Not sure why they did things this way, but we're stuck with it now. Maybe we can solve it with a quirk in the USB-HID driver? Identify mac keyboards and switch the keycodes at that level? Would that work?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>