Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > I think we should use Command in menus as well. At this point it's a pile of mess. If we are going to standardise it, we should do it our way. Otherwise people will continue to be confused, at least this way they'll learn it once and never forget it again. They will get used to it.
> > 
> > No users won’t get used to, they won’t look it up in the user guide, they won’t look at Keymap preferences, they will simply be forever confused. We cannot ignore human nature when we design our HCI, that is folly.
> 
> Every other OS present some sort of welcome wizard on first run (unless it's some obscure Linux distribution or BSD), and we should do this as well. Important details like these should be conveyed with nice big key illustrations and possibly a small interactive tutor. Avoiding this won't make the confusion go away.

Haiku already provides a "wizard" on first run, called FirstRunBootPrompt, and it provides an option to select your keymap with illustrations. We must fix the (Mac) keymap to swap the keys that are assigned to the command and option roles... and I thought that we were already doing that! Maybe someday we'll be able to detect if you are using an Apple or PC keyboard and do better than this. Unfortunately if we did detect the keyboard as Apple or PC, we'd be the first OS to do so, because every other OS has punted on this problem in one way or another -- be it Windows, macOS or Linux.

> Patch Set 1:
> 
> > No users won’t get used to, they won’t look it up in the user guide, they won’t look at Keymap preferences, they will simply be forever confused. We cannot ignore human nature when we design our HCI, that is folly.
> 
> If we solve this problem with AltGr I think we can manage to show the actual key names to users (control / alt / windows / altgr on pc; control/command/option on mac keyboards), actually matching their keyboards, in menus. So only the keymap preferences will remain a bit confusing by talking about command and option in all cases. But we can revisit (again) how we handle that.
> 
> There is the additional problem of having to handle cases where the PC and MAc keyboards are different (as is the case for French), and the cases where they are identical except for the swapping of command and option (as is the case for US).
> 
> But yes, Apple did mess things up when they designed their USB keyboards, I guess? Not sure why they did things this way, but we're stuck with it now. Maybe we can solve it with a quirk in the USB-HID driver? Identify mac keyboards and switch the keycodes at that level? Would that work?

Nah, we have to fix it up in the keymaps for now... maybe someday we can do better but right now we can't because our keyboard API does not support multiple keyboards.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>