Update patch set 1
Patch Set 1:

(1 comment)

In general I wonder if there is much use for this outside Tracker? You can't really do much with a node_ref on its own, you almost always want the parent directory somehow (then you get an entry_ref).

Tracker is a bit of a special case here because its main job is to manage directories, and having a copy of the path in all entry_refs would be very inefficient. But I suspect few other apps will need it.

Well, I guess we never know and it can't hurt to have it available anyway.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>