Update patch set 16
Patch Set 16:

> You are right that there is not a meaningful difference unless you compare the two side by side, although the thumbnail quality isn’t very good in the first place / could be a lot better so it’s hard to say. If we had a better downscaling algorithm available from BView::DrawBitmap() that would make a bigger difference.

Does it not already handle bilinear filtering? I think there is a flag for that somewhere (possibly an undocumented API). Or do you need something even better?

> Lots of people view files using only 1 or 2 icon sizes, hardly anyone uses >64x64 except on the Desktop. I don’t have enough data to prove or disprove this, and I have no idea how to get it, so I’m going off intuition here. People expect to take the thumbnail generation hit when switching icon sizes, I don’t see a problem here. Using bigger icon sizes means more “waste” one way or the other, at least with my way the extra waste is only incurred when and if you actually use those sizes.

You have to remember one thing here, with thumbnails generation available people *will* start to use the larger font size in their picture folders. Until now it was useless to have large icons, but large thumbnails are a completely different thing, and having them does increase usability a lot.

In fact, now I'm starting to wonder if thumbnails at smaller sizes are useful at all. At 16x16 it would look like a mashup of pixels, probably, unless the image is already a small one. Even at 32x32 it's maybe not that useful, especially if the image is in a non-square aspect ratio, the picture may end up being 24x32px (unless we crop the picture sides instead?). Maybe it would make sense to skip these altogether, and use thumbnails only at a bit larger zoom levels? But then it reduces the discoverability of the feature. So maybe show the thumbnails, even if they are unreadable, just to hint the user that they can zoom in and actually see something useful.

Patch-set: 16