Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> In general I wonder if there is much use for this outside Tracker? You can't really do much with a node_ref on its own, you almost always want the parent directory somehow (then you get an entry_ref).

This feature may never be used outside of Tracker, but it could be used by Tracker in a couple of places.

> Tracker is a bit of a special case here because its main job is to manage directories, and having a copy of the path in all entry_refs would be very inefficient. But I suspect few other apps will need it.
> 
> Well, I guess we never know and it can't hurt to have it available anyway.

Node refs are used for both files and directories in Tracker... but there's no way to get the parent directory of a file from its node ref. There is nothing special about how Tracker is using node refs, they may be used any time you want to pass a reference to a file and don't care about where that file is in the file system hierarchy.

I wonder why Be didn't include this in BMessage already? Probably because they didn't care enough to add it since you can pass the device id and inode separately instead, but this is not as convenient.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>