Update patch set 16
Patch Set 16:

> > I also made a comparison of 40x40 downscaled from 64x64 instead of 128x128, but there is essentially no difference in thumbnail image quality in this case, so you'll just have to imagine what it looks like (both sides identical.) Therefore, from this I determine that downscaling from 64x64 to 40x40 is fine, but downscaling from 128x128 to 40x40 is not. Agreed?

The downscaling algorithm can always be improved later. And yes, for it to look decent we really need the bilinear filtering, especially at such small sizes. So we have to get that working with transparency. But we can merge a first version of this, and then switch the algorithm later when we have fixed the bilinear scaling to handle transparency, right?

Patch-set: 16