Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think they have some sort of "hybrid approach" at this point, so I guess we should see precisely what they are doing and try to emulate that.

They use the Linux code with a wrapper because there is no other non-crazy way to do this. Why would they make large changes to the driver when it's much easier to just have some .h files for wrapping it up?

There can be some hybrid approach, sure, but what's in this and the next commit is the bare minimum to get anything to compile at all.

Also, Dragonfly seems to be based on Linux 4.12, of which the last release was in 2017. I guess this shows the amount of work required to keep things going. If we go that way, the less change we make to the Linux code, the better. Reusing too much code from one of the *BSD means we have to wait for them to update, and only then we can get their changes. So, we end up with several years of lag behind Linux. What's the point then?

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>