Update patch set 2
Patch Set 2:

IIRC, Linux driver QC, even for video drivers, is pretty bad; so often they do not update partially because critical hardware is broken, and also because the DRM code changes so quickly that it is not so easy to keep a compatibility layer working.

If we are going to go the 100%-compat-layer route, then FreeBSD is the better template, I think.

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>