Update patch set 2
Patch Set 2:

> 1) Haiku's video card drivers need to expose DRM devices in /dev/graphics/*
> 2) Haiku's video card drivers need the DRM drivers integrated into them, or we need to wrap each DRM driver with our graphics card API

> Keep in mind both steps above involve the video card drivers leveraging DRM compatibility headers/libraries from Linux

I'd say it's either we use everything from the Linux driver, or everything from our own code. I don't think it's possible to split apart the modesetting from the other parts of the Linux drivers, at least for the Intel one (the others I don't know, didn't have to look at them).

So that's our choice: try to port the Linux drivers, and replace ours, which requires a lot of work on the compatibility layer, or, try to implement drm on our own, which also needs quite some code (in addition to keeping the drivers themselves up to date).

At this point I don't know which is easier to get working and to keep up to date. Possibly the only way to know, is to try this, and see how far we get into at least trying to compile the drivers.

> Linux video card drivers are also GPL, so a case exists here to keep them at an arms length outside of our codebase. (aka, haikuports package with mesa + libGL)

If the drivers are GPL that's a no-go from me. But at least the Intel one isn't, it's MIT. On the other hand we have to be careful when we write the compatibility layer as there is code from Linux we can't import.

> As we work on DRM + code + headers and better determine the requirements, we can then begin thinking about rolling DRM into Haiku more tightly.

Well, big projects always work one step at a time. For those of you who have not followed since the beginning, here are the steps acomplished so far:

- Using Mesa in software rendering mode: too long ago for me to remember
- Rudolfc's Utahglx experimental acceleration with old nvidia cards (http://rudolfs-place.nl/BeOS/NVdriver/nvidia3d.html) around 2008
- Early efforts on porting Gallium3D with help from AROS developers in 2009 by aljen: http://revolf.free.fr/FOSDEM/2010/FOSDEM2010_ALTOS_007_Gallium3D.pdf
- Work by kallisti5 to upstream our API into Mesa, and get modern Mesa version ported to Haiku (2011: https://www.mail-archive.com/mesa-dev@lists.freedesktop.org/msg15887.html)
- kallisti5's librendomatic: 2014-2016
- This work on Linux compatibility: hamishm, 2015

So it's actually an ongoing project with some progress over the last 10 years.

On to the next thing :)

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>