Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4: Code-Review+2
> > 
> > IIRC the Get* and Has* methods have been deprecated for a long time (possibly already in BeOS?), the expectation is than you use Find* and check the error code. This avoids searching the BMessage twice (to check if the data is there, then to actually get it).
> 
> The Get* methods were added by us (Axel in hrev44739) so they aren't deprecated. The Get* methods allow you to set a default value to use in the case that the Get* methods fails. The Has* methods however, have been there a long time. Maybe it's a bad idea to use these because it forces you to search the BMessage twice, but, I still don't see why that makes them deprecated and unworthy of being documented. If these methods really are discouraged from use, perhaps I should not include HasNodeRef() in the class.

Oops, yes, I meant the versions of Find* that return the value directly. The Be Book says about these: "(Several functions, such as FindRect() and FindInt32(), have versions that return the found value directly. These versions don't report errors and may not be supported in the future.)"

And the Has* functions are not documented in R5 Be Book, but they are in https://github.com/jscipione/r5headers/blob/master/headers/be/app/Message.h under a "Private, reserved, or obsolete" comment.

The R3 Be Book (http://victor3d.karelia.pro/BeOS/BeBookR3/The%20Application%20Kit/Message.html) documents the "Has*" functions as: "If the message contains the data you're looking for, you probably will want to call Find...() to get it. Since the Find...() functions return errors if the message doesn't have data of the requested name and type, it's more efficient to just call Find...() and not bother with Has...()."

So, rather useless in most cases, and costly because you end up searching twice. And so Be deprecated them and we didn't change that.

As for adding them for new types, I don't know, I think it makes sense to keep all types working identically, so I'd still add it.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>