Update patch set 2
Patch Set 2:

(3 comments)

Thanks for your work on improving the user experience! Not the easiest thing (to get good understanding and agreement on what is actually wanted) but important to keep the polished feeling of Haiku.

There are some minor comments on coding style and one on behavior change below. Other than that, looks good!

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>