Update patch set 2
Patch Set 2:

Oddly, the package files are kind of like tar files, except compressed and exposed as a file system.  Is it that much slower to use icons as files from a package than from a tar file?  Somewhat more awkward for other programs or people to access.  Though if this is purely internal to HDS, and there aren't any other GUI interfaces to the repositories, tar is fine.

Reminds me of the data files we used in making games, same idea.  Some game engines would expand the whole data file into memory when loading a level (those sorts didn't let you allocate new objects, so if you were juggling and had at most 10 balls, you'd have to have 10 of them in the data file, can't make a copy of just one).

Patch-set: 2
CC: Gerrit User 1000038 <1000038@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>