Update patch set 2
Patch Set 2:

Is it really worth changing this though? There is a ton of documentation out there (on websites, forums, mailing lists) for the current behaviour.

And I could argue that it is not an improvement that you now need two separate options with arguments to accomplish the same goal. 

Arguably, if you would like to separate out the source dir argument, you could, but then arguably it would interplay with the target-arch argument.

So:
./configure --cross-tools-source ../buildtools/ --target-arch x86_gcc2 --target-arch x86

Patch-set: 2
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>