Update patch set 1
Patch Set 1:

> I've had a look at Linux's drivers and, from what I can gather, intel_pstates shifts the value and acpi_cpufreq doesn't. So this may depend on the model, don't merge without confirmation from tests in other hardware.

It seems the register value is indeed CPU model specific and we shouldn't be writing arbitrary values there. I found an osdev.org forum thread where it was reverse engineered for some models, the high bytes there sets the CPU frequency, while the low byte sets the voltage.

Apparently we should be using the ACPI _PSS table to convert an arbitrary "pstate" value (0 = max powersaving, 255 = full speed for example) into something appropriate for writing in the register. If we don't do that, we have to make sure what we're writing here for different CPU models.

Patch-set: 1