Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > obviously this changes the size of the cpu_info struct, which breaks users of get_cpu_info.
> > get_cpu_info isn't implemented with the size of the struct, so it can't get extended. I didn't feel that adding a syscall for this would be wise. Opinions?
> 
> Maybe we could add get_cpu_info_etc() that includes the size, and introduce a new struct e.g. extended_cpu_info to allow for forwards compat?

We can keep the new struct as cpu_info and rename the old one (since this is C ABI, the naming isn't reflected in the symbol names). A similar approach was used for get_system_info when removing the hardcoded B_MAX_CPU_COUNT from it. The old struct and symbols are not in the public headers anymore, but are still callable by already compiled applications.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>