Update patch set 4
Patch Set 4:

That's still asking too many questions.

The choice of EFI vs GPT doesn't need to be there. Either always use GPT; or use GPT for EFI systems and MBR for BIOS systems. And ideally we should detect how the OS is booted and select the rigth mode by default.

Also I think this should be in Installer, not in DriveSetup. Having to start multiple apps is already enough to confuse users. The flow should be something like:

FirstBootPrompt -> Installer "select drive list" -> Select "wipe whole drive X and install there" -> Install -> Reboot, done. No need to click a "edit partitions" thing or anything like that.

Is that too ambitious for the next release? Are we going to write all this UI knowing that eventually it is still too complicated?

Patch-set: 4
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>