Update patch set 6
Patch Set 6:

> As can be seen in my comparison between the BeOS R5 Backgrounds Preferences and the same binary running on Haiku [0], we want the label to render a bit higher up than the pop up menu for it to match BeOS, which we do for compatibility, and because it looks nicer IMHO.
> 
> [0] https://insightfactory.tumblr.com/post/623900047552200704/beos-r5-backgrounds-preflet-on-haiku-and-on-beos

My complaint is that the patch is not complete. You have moved the text label up, but not when it is a custom view. As a result, for example in our background preferences, the "Preview" and "Current workspace" texts are not aligned with each other (you can look at the "e" char in each other, maybe setting a larger font size to make it more obvious).

The fLabel custom view, when it's there, should be shifted up similarly. Then I'm happy to merge this :)

Patch-set: 6