Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > As can be seen in my comparison between the BeOS R5 Backgrounds Preferences and the same binary running on Haiku [0], we want the label to render a bit higher up than the pop up menu for it to match BeOS, which we do for compatibility, and because it looks nicer IMHO.
> > 
> > [0] https://insightfactory.tumblr.com/post/623900047552200704/beos-r5-backgrounds-preflet-on-haiku-and-on-beos
> 
> My complaint is that the patch is not complete. You have moved the text label up, but not when it is a custom view. As a result, for example in our background preferences, the "Preview" and "Current workspace" texts are not aligned with each other (you can look at the "e" char in each other, maybe setting a larger font size to make it more obvious).
> 
> The fLabel custom view, when it's there, should be shifted up similarly. Then I'm happy to merge this :)

You are missing that the placement difference between the box text and the menu field text is intentional in order to reproduce what BeOS was doing. If you take a screen ruler put it over the text on the linked screen shot you'll see that the baseline of the text "Preview" and the menu field label "Current Workspace" are at exactly the same vertical position. Here is a comparison I made to show the baseline:

https://64.media.tumblr.com/1398f017a5a11fcea8956e6bb5c445b4/2b55696d07312a3b-93/s540x810/84678a650e97f03f10ee94ee300284472a55aa3f.png

There is a trick of the eye that they look like they are in different positions, but you can see that the baselines line up exactly. So what you are asking for not what I'm trying to do, the patch is complete, but I'm trying to achieve a different result than you are expecting.

Note that this comparison shows the same BeOS R5 Backgrounds binary running on BeOS R5 and Haiku, and that the Haiku Backgrounds preference dialog is a bit different, it has a bold BBox label for instance, but to make the comparison the same I ran the BeOS binary on Haiku.

If I'm missing something here I apologize, maybe I don't get what you are saying or you are seeing a different result.

Patch-set: 6
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>