Update patch set 6
Patch Set 6:

> You are missing that the placement difference between the box text and the menu field text is intentional in order to reproduce what BeOS was doing. If you take a screen ruler put it over the text on the linked screen shot you'll see that the baseline of the text "Preview" and the menu field label "Current Workspace" are at exactly the same vertical position. Here is a comparison I made to show the baseline:

No, I checked that (with magnify) and it's not exactly correct at all font sizes. I used a very large font size (I think 30) and there was a difference of 4 pixels. Maybe it's just a small rounding error that gets worse with larger fonts.

> There is a trick of the eye that they look like they are in different positions, but you can see that the baselines line up exactly. So what you are asking for not what I'm trying to do, the patch is complete, but I'm trying to achieve a different result than you are expecting.

We are expecting the same thing, but your patch does not work at all font sizes.

> Note that this comparison shows the same BeOS R5 Backgrounds binary running on BeOS R5 and Haiku, and that the Haiku Backgrounds preference dialog is a bit different, it has a bold BBox label for instance, but to make the comparison the same I ran the BeOS binary on Haiku.

I made the comparison using the Haiku one, but changing my bold font in preferences to be the same as the normal font.

Patch-set: 6