pkgman: add dist-upgrade command to upgrade from medium
The dist-upgrade command will allow a user to perform a full sync of their
system based of a medium. The command does some basic checks to determine
whether a path points to a medium (or to something that resembles an
installation medium), run a full-sync command, and copy over the repository
config files of the installation medium.

WIP: There should be a check to make sure that the install medium does not
contain an older version of Haiku. It won't matter for the upgrade itself,
but it may be destructive to the repository configuration files.