pkgman: add medium-upgrade command to upgrade from medium
The medium-upgrade command will allow a user to perform a full sync of their
system based of a medium. The command does some basic checks to determine
whether a path points to a medium (or to something that resembles an
installation medium), run a full-sync command, and copy over the repository
config files of the installation medium.

Change-Id: Iec985a8380374b6798f5bfda83b12e38b241a9dd