Update patch set 2
Patch Set 2:

> How do we test this change and get it past the post?

It's not just about testing. The technicql reasons for not merging it yet have been lost in the discussion it seems.

The problem in the current version of the code is that it is not backwards compatible. If someone has a blacklist file in place (as documented in https://www.haiku-os.org/guides/daily-tasks/blacklist-packages/), it will not work anymore because the "EntryBlacklist" will not be recognized.

So, we should make this code look for either "EntryBlqcklist" or "BlockedEntries" in this file. This way people having a file in the old format will not see their files suddenly unblocked when updating.

Then for the testing:
- Block some files using the old format
- Update haiku to include this patch
- Check they are still blocked
- Block the files now using the new format
- Check they are still blocked.

Patch-set: 2