Use all the fallback fonts
When drawing text you may need a glyph that is not in the current font.
Instead of not drawing it at all or use a placeholder, some system fonts
are tried first and used as a fallback font if the glyph is found there.

Only one font is used as a fallback for all the string run, though, so if
it needs glyphs from two different system fonts, you'll only get the ones
in the first.

This problem is compounded when the application uses different runs for the
same text, for example when redrawing for selection or cacheing positions
while laying out the text.

So let's use the whole list of fallbacks, in their preferred order, whenever
a glyph is not found in the original font, including them whether they have
that specific glyph or not. That way the next unavailable glyph will also
be searched in all the fallbacks in that same order and not only in the one
used for the first unavailable glyph.

Solves #16404, #16213, #16221, maybe other garbled-text-in-web issues.

Does not solve similar issues of getting slightly different positions and
widths for a whole string than for the sum of its parts due to subpixel
rendering (cf #15007 comment 2).

Change-Id: I40664e32c8ecd5f09e10aa7a60fcdaec230471b6