Update patch set 2
Patch Set 2:

I don't follow you here.

The _WriteLockAndAcquireFallbackEntry function is called for each character and will find the correct font for it, or at least that was the idea. But you removed the charCode parameter to it and now it is not able anymore to locate the correct font for a given character. And I think you moved the code doing that elsewhere.

It seems to make the code a lot more complicated (I saw an *** pointer? That's crazy) and even harder to follow than it already was (not that the existing code is great, it was not so easy to integrate this with the existing font cache design).

What makes you think a single fallback is used for the whole run, when _WriteLockAndAcquireFallbackEntry is called for each missing character?

I would prefer that we simply fix the bug while keeping the way it is done currently (one call to _WriteLockAndAcquireFallbackEntry for each missing character), is that not possible?

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>