Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What makes you think a single fallback is used for the whole run, when _WriteLockAndAcquireFallbackEntry is called for each missing character?
> 

It is not. It is called for the first missing character, then the fallback font and writeLock are set if a font is found with that character and thus _WriteLockAndAcquireFallbackEntry is never called again.

You are right in the other points, of course. Hey, I don't even understand references and it took me a whole day just to get this to a point where it compiled. The *** is certainly a sign of something sketchy, and I'm sure a new here and a malloc there is also not a good sign. I'm definitely not asking for this or even something similar to be merged, but I think it will have been worth the effort if someone who can code gets the problem and finds a real solution.

Patch-set: 2
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>