Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > The _WriteLockAndAcquireFallbackEntry function is called for each character and will find the correct font for it, or at least that was the idea.
> 
> I did try calling it for each missing character back when I found about this, but always got deadlocks, crashes or whatever, not even getting to Tracker.
> 
> You've made me try again, and I may have learned something since those first tries, because this time it seems to work, I don't even get the deadlocks I got with the previous patch while playing with FontBoy (at least I haven't yet). I'm assuming the fallbackCacheReference can be unset just after getting the glyph, I don't know the performance implications and I will never be sure about all the locking, but this has just changed from root cause proof of concept to a real patch for review.

That's better, but not perfect yet :)

From what I understand/remember about the locking.

The original idea is that we start with just a read lock on the font cache. This means we can get cached characters from it, but not add new ones there.

Whenever we find that a character is not in the cache (this can happen because the character was never drawn yet), we switch to a write-lock instead. This allows us to add new things to the cache.

The advantage of this is that many strings can be drawn in different threads as long as they don't need the write lock (multiple readers is no problem). And after some time, it's likely that all used parts of the font will be in the cache and no write lock at all will be needed.

Currently, we do this at the same time as getting the fallback entry. But with the current version of the patch we are now doing it multiple time, which should not be needed (once we get the write lock, we can as well keep it until we're done drawing the whole string).

So we need to split this function in two, or maybe simply make it check if we are already in write lock, and do just the font fallback part in that case. (splitting in two would be clearer, but I don't know if it's possible, I think I remember there is more complex locking going on inside the function that's not so easy to change).

Patch-set: 3