Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Whenever we find that a character is not in the cache (this can happen because the character was never drawn yet)

Oh! I hadn't noticed this. So it's even worse than I thought (for unlikely cases?): when you use a new font you go for a fallback (and write-lock it) with the very first character whether you need that fallback or not. And you don't get a new one if a later character really needs it. And you get different results if you repeat exactly the same string with the same font but enough glyphs have been cached.

> [more insight into the locks]
> 
> Currently, we [write-lock the font] at the same time as getting the fallback entry. But with the current version of the patch we are now doing it multiple times, which should not be needed (once we get the write lock, we can as well keep it until we're done drawing the whole string).

Not really, if I'm reading it correctly.

The first thing _WriteLockAndAcquireFallbackEntry does is unlock the main font, apparently because it later locks the FontManager and that would obviously (certainly not for someone who doesn't know how this works) deadlock. It doesn't unlock the fallback because there is none, and that's the one I'm unlocking (and recycling, which may be unnecessary right then).

I guess we should first check if our font can create the glyph and in that case make sure we have a write lock and call CreateGlyph without fallback. But if our font can't create the glyph, we have to look for a fallback (the one we may have may not be the preferred one even if it can create the glyph) and it seems we have to release the locks there, so we may as well release the fallback one as soon as we use it: we won't ever need it again without releasing it first.

In fact, _WriteLockAndAcquireFallbackEntry should be _FindMeAGlyphSomewehere and return the glyph, with LayoutGlyphs having nothing to do with fallbacks or even knowing if the main font is in read or write lock.

There may be another problem: isn't the font locking in _WriteLockAndAcquireFallbackEntry a recipe for a deadlock? Say two threads are using different fonts both of which are in the fallback list. They need to go for a fallback and find the font the other thread is using as a suitable one. Those fonts were unlocked at the beginning of _WriteLockAndAcquireFallbackEntry, so each thread write-locks the main font the other one is using, they go out of the loop, try to write-lock their own font and wait forever for the other thread to release it. And anyone who later needs a non cached glyph from any of those fonts also blocks.

Even if it is not a problem now it may be more likely to happen if we start looking for fallback fonts every time we need them, and I don't have a solution for this.

Patch-set: 3
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>