Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Understood.   The issue here is golang passes NULL with a length of 0 to write. This "seems to work" on all other platforms.
> 
> 
> 1)   If length > 0, buffer != NULL
> 2)   AND is not a user address (or NULL)
> 3)   B_BAD_ADDRESS.
> (later on)
> 4)   if length == 0, return.
> 
> 
> Thus.  The NULL check is only "bypassed" on on condition 1.
> Condition 1 is then "returned" due to the length of 0 at change line 4.

That assumes length == 0 was passed in though. If the passed in length is > 0 and the buffer is NULL, the user address check will be bypassed, which is what Axel and I were trying to get at. Under that circumstance, unless I'm misreading the code, the invalid buffer will happily get passed along to the fd_read/write ops. 

The golden rule is, do not trust the user to be passing valid inputs to syscalls. I realize the case you're trying to get working, but the way you're currently trying to rectify it opens up other error conditions that weren't previously present.

Patch-set: 4